tools/seasonality_analyzer/requirements.txt
streamlit>=1.30.0
yfinance>=0.2.36
pandas>=2.1.4
numpy>=1.26.3
plotly>=5.18.0
scipy>=1.11.4
python-dateutil>=2.8.2
